postman test

+++ ewm-service/src/main/java/ru/practicum/category/CategoryServiceImpl.java
@@ -1,15 +1,13 @@
 package ru.practicum.category;
 
-import jakarta.validation.ConstraintViolationException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.event.EventRepository;
-import ru.practicum.exception.AlreadyExistsException;
-import ru.practicum.exception.NotFoundException;
-import ru.practicum.exception.ValidationException;
+import ru.practicum.exception.*;
 
 import java.util.List;
 
@@ -18,36 +16,41 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryMapper mapper;
     private final CategoryRepository repository;
-    private EventRepository eventRepository;
+    private final EventRepository eventRepository;
 
     public CategoryDto addCategory(NewCategoryRequest request) {
         try {
+            findByName(request.getName());
             Category category = new Category();
             category.setName(request.getName());
             return mapper.categoryToCategoryDto(repository.save(category));
-        } catch (ConstraintViolationException e) {
+        } catch (DataIntegrityViolationException e) {
             throw isUnique(e, request.getName());
         }
     }
 
     public void deleteCategory(long catId) {
         findCategoryById(catId);
-//        if (!eventRepository.findByCategoryId(catId).isEmpty()) {
-//            throw new
-//        }
+        if (!eventRepository.findByCategoryId(catId).isEmpty()) {
+            throw new CategoryHasEventsException("Category has events");
+        }
         repository.deleteById(catId);
     }
 
     public CategoryDto modifyCategory(long catId, NewCategoryRequest request) {
         try {
+
             Category category = findCategoryById(catId);
+            if (!category.getName().equals(request.getName())) {
+                findByName(request.getName());
+            }
 
             if (request.getName().length() > 50) {
                 throw new ValidationException("Name length can not be > 50");
             }
             category.setName(request.getName());
             return mapper.categoryToCategoryDto(repository.save(category));
-        } catch (ConstraintViolationException e) {
+        } catch (DataIntegrityViolationException e) {
             throw isUnique(e, request.getName());
         }
 
@@ -66,7 +69,8 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     private Category findCategoryById(long catId) {
-        return repository.findById(catId).orElseThrow(() -> new NotFoundException("Category with id=" + catId + " nto found"));
+        return repository.findById(catId)
+                .orElseThrow(() -> new NotFoundException("Category with id=" + catId + " nto found"));
     }
 
     private RuntimeException isUnique(RuntimeException e, String categoryName) {
@@ -74,5 +78,11 @@ public class CategoryServiceImpl implements CategoryService {
             return new AlreadyExistsException("Category with name " + categoryName + " already exists");
         }
         return new ValidationException("Field name is not correct");
+    }
+
+    private void findByName(String name) {
+        if (repository.existsByName(name)) {
+            throw new DuplicatedDataException("Name already in use");
+        }
     }
 }
